move Icon struct to common mod
